Add license
Dual-licenses the software under the Apache-2.0 and BSD-3-Clause
licenses and assigns the copyright to Caltech.

This is consistent with the open source plan for JPL-developed software
under the OPERA project -- I'm not sure if the contract for this work
specifies that a different license/copyright should be used, so feel
free to correct me.